Building module of oil and hydraulic changes and creating a method for avoid duplicated images

+++ src/main/java/com/app/usochicamochabackend/review/application/dto/InspectionFormRequest.java
@@ -7,7 +7,7 @@ public record InspectionFormRequest(
         String UUID,
         Boolean isUnexpected,
         LocalDateTime dateStamp,
-        BigInteger hourMeter,
+        Integer hourMeter,
         String leakStatus,
         String brakeStatus,
         String beltsPulleysStatus,
