+++ src/main/java/com/app/usochicamochabackend/review/application/dto/InspectionResponse.java
@@ -13,7 +13,7 @@ public record InspectionResponse(
         String UUID,
         Boolean isUnexpected,
         LocalDateTime dateStamp,
-        BigInteger hourMeter,
+        Integer hourMeter,
         String leakStatus,
         String brakeStatus,
         String beltsPulleysStatus,
